+++ CHANGELOG.md
@@ -49,6 +49,9 @@
 * Added `--conda-package-version` flag for specifying version of conda package in `nf-core modules create`. ([#1238](https://github.com/nf-core/tools/issues/1238))
 * Add option of writing diffs to file in `nf-core modules update` using either interactive prompts or the new `--diff-file` flag.
 * Fixed edge case where module names that were substrings of other modules caused both to be installed ([#1380](https://github.com/nf-core/tools/issues/1380))
+* Tweak handling of empty files when generating the test YAML ([#1376](https://github.com/nf-core/tools/issues/1376))
+    * Fail linting if a md5sum for an empty file is found (instead of a warning)
+    * Don't skip the md5 when generating a test file if an empty file is found (so that linting fails and can be manually checked)
 
 ## [v2.2 - Lead Liger](https://github.com/nf-core/tools/releases/tag/2.2) - [2021-12-14]